+++ java/com/turbovnc/rfb/HeaderParameter.java
@@ -1,4 +1,4 @@
-/* Copyright (C) 2017 D. R. Commander.  All Rights Reserved.
+/* Copyright (C) 2017-2018 D. R. Commander.  All Rights Reserved.
  *
  * This is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -28,6 +28,10 @@ public class HeaderParameter extends VoidParameter {
 
   public final boolean setParam(String v) {
     throw new ErrorException("Cannot set header parameter");
+  }
+
+  public final void reset() {
+    throw new ErrorException("Cannot reset header parameter");
   }
 
   public final String getDefaultStr() {