+++ java/com/turbovnc/rfb/IntParameter.java
@@ -1,5 +1,5 @@
 /* Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
- * Copyright (C) 2012, 2017 D. R. Commander.  All Rights Reserved.
+ * Copyright (C) 2012, 2017-2018 D. R. Commander.  All Rights Reserved.
  *
  * This is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -56,6 +56,8 @@ public class IntParameter extends VoidParameter {
     value = v;
     return true;
   }
+
+  public synchronized void reset() { value = defValue; }
 
   public String getDefaultStr() {
     if (defValue >= 0)