+++ java/com/turbovnc/rfb/VoidParameter.java
@@ -1,5 +1,5 @@
 /* Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
- * Copyright (C) 2012, 2017 D. R. Commander.  All Rights Reserved.
+ * Copyright (C) 2012, 2017-2018 D. R. Commander.  All Rights Reserved.
  *
  * This is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -35,6 +35,9 @@ public abstract class VoidParameter {
 
   public abstract boolean setParam(String value);
   public boolean setParam() { return false; }
+
+  public abstract void reset();
+
   public abstract String getDefaultStr();
   public abstract String getValueStr();
   public abstract String getValues();